Add files via upload
In order to enhance our driving experience on Computer Devices (in games or any other driving simulator tool)
I have created a Virtually handling the controls of that uses hand movement to capture and take inputs for the simulation.
The webcam detects the coordination of the hand and based of the coordination it directs the vehicle in the simulator.

We need to install the Mediapipe and opencv for image processing and video capturing, you can install them using pip install. 

In this a framework for human hand movement recognition is implemented for users to use/play a driving simulator/game with more immersion and accuracy using a webcam. 
Here the user shall make hand gesture in a way such that you are using a real car’s steering wheel. Our module will use the webcam to capture the hand’s/fist’s coordinates by calculating the slope whenever the user’s hand gesture gets tilted either left or right, the vehicle in simulator/game will turn accordingly.